+++ packages/react/index.ts
@@ -1,10 +1,12 @@
-const { isPackageExists } = require('local-pkg')
+import type ESLint from 'eslint'
+import { isPackageExists } from 'local-pkg'
+
 const tsExits = isPackageExists('typescript')
 
 if (!tsExits)
 	console.warn('[@eslint-sets/eslint-config] TypeScript is not installed, fallback to JS only.')
 
-const config = {
+const config: ESLint.Linter.BaseConfig = {
 	plugins: [
 		'react',
 		'react-hooks',